+++ assets/src/scripts/monitoring-location/components/hydrograph/discrete-data.test.js
@@ -1,8 +1,8 @@
 import {select} from 'd3-selection';
 import {scaleLinear} from 'd3-scale';
 
-import {circleMarker} from 'd3render/markers';
-import {drawGroundwaterLevels, getGroundwaterLevelsMarker} from './discrete-data';
+import {drawGroundwaterLevels} from './discrete-data';
+
 
 describe('monitoring-location/components/hydrograph/discrete-data', () => {
     describe('drawGroundwaterLevels', () => {
@@ -12,10 +12,12 @@ describe('monitoring-location/components/hydrograph/discrete-data', () => {
         beforeEach(() => {
             svg = select('body').append('svg');
             gwLevels = [
-                {value: '14.0', dateTime: 1491055200000},
-                {value: '14.5', dateTime: 1490882400000},
-                {value: '13.0', dateTime: 1490536800000},
-                {value: '12.0', dateTime: 1489672800000}
+                {value: '14.0', dateTime: 1491055200000, label: 'Approved', classes: ['gw-class', 'approved'], radius: 5},
+                {value: '14.5', dateTime: 1490882400000, label: 'Provisional', classes: ['gw-class', 'provisional'], radius: 5},
+                {value: '13.0', dateTime: 1490536800000, label: 'Revised', classes: ['gw-class', 'revised'], radius: 5},
+                {value: '12.0', dateTime: 1489672800000, label: 'Provisional', classes: ['gw-class', 'provisional'], radius: 5},
+                {value: '11.0', dateTime: 1489672300000, label: 'Provisional', classes: ['gw-class', 'provisional'], radius: 5},
+                {value: '13.0', dateTime: 1489672100000, label: 'Provisional', classes: ['gw-class', 'provisional'], radius: 5}
             ];
             xScale = scaleLinear()
                 .domain([0, 100])
@@ -29,16 +31,19 @@ describe('monitoring-location/components/hydrograph/discrete-data', () => {
             svg.remove();
         });
 
-        it('Renders 4 circles for each gw level', () => {
+        it('Renders correct number of circles with correct class for each gw level', () => {
             drawGroundwaterLevels(svg, {
                 levels: gwLevels,
                 xScale: xScale,
                 yScale: yScale
             });
-            expect(svg.selectAll('circle').size()).toBe(4);
+            expect(svg.selectAll('circle').size()).toBe(6);
+            expect(svg.selectAll('.approved').size()).toBe(1);
+            expect(svg.selectAll('.provisional').size()).toBe(4);
+            expect(svg.selectAll('.revised').size()).toBe(1);
         });
 
-        it('A second call to render with no gw levels renders no circles', () => {
+        it('A second call to render with no gw points renders no circles', () => {
             drawGroundwaterLevels(svg, {
                 levels: gwLevels,
                 xScale: xScale,
@@ -50,12 +55,6 @@ describe('monitoring-location/components/hydrograph/discrete-data', () => {
                 yScale: yScale
             });
             expect(svg.selectAll('circle').size()).toBe(0);
-        });
-
-        describe('getGroundwaterLevelsMarker', () => {
-            it('Expects to return a circle marker', () => {
-                expect(getGroundwaterLevelsMarker().type).toBe(circleMarker);
-            });
         });
     });
 });