+++ assets/src/scripts/monitoring-location/components/hydrograph/legend.test.js
@@ -46,8 +46,8 @@ describe('monitoring-location/components/hydrograph/legend module', () => {
         });
 
 
-        it('Should have 6 legend markers', () => {
-            expect(selectAll('.legend g').size()).toBe(6);
+        it('Should have the correct number of legend markers', () => {
+            expect(selectAll('.legend g').size()).toBe(9);
         });
     });
 });