+++ assets/src/scripts/monitoring-location/components/hydrograph/selectors/discrete-data.js
@@ -1,17 +1,39 @@
+import isEqual from 'lodash/isEqual';
+import uniqWith from 'lodash/uniqWith';
 import {DateTime} from 'luxon';
 import {createSelector} from 'reselect';
 
 import config from 'ui/config';
 
-import {getGroundwaterLevels} from 'ml/selectors/hydrograph-data-selector';
+import {getGroundwaterLevels, getPrimaryParameter} from 'ml/selectors/hydrograph-data-selector';
+
+const GW_LEVEL_RADIUS = 7;
+const GW_LEVEL_CLASS = 'gw-level-point';
+
+const APPROVAL_QUALIFIERS = {
+    'A': {label: 'Approved', class: 'approved'},
+    'P': {label: 'Provisional', class: 'provisional'},
+    'R': {label: 'Revised', class: 'revised'}
+};
+
+const getDetailsForApprovalCode = function(qualifiers) {
+    const approvalQualifier = qualifiers.find(qualifier => qualifier in APPROVAL_QUALIFIERS);
+    if (approvalQualifier) {
+        return APPROVAL_QUALIFIERS[approvalQualifier];
+    } else {
+        return APPROVAL_QUALIFIERS['P'];
+    }
+};
 
 /*
  * Returns a selector function that returns the groundwater levels that will be visible
- * on the hydrograpnh
+ * on the hydrograph
  * @return {Function} which returns an {Array} of groundwater level object with properties:
  *      @prop {Float} value
- *      @prop {Array of String} qualifiers
  *      @prop {Number} dateTime
+ *      @prop {Array of String} classes - a class that can be used to style this point
+ *      @prop {String} label - a human readable label for this kind of point
+ *      @prop {Number} radius - used to draw the circle marker
  */
 export const getGroundwaterLevelPoints = createSelector(
     getGroundwaterLevels,
@@ -20,29 +42,38 @@ export const getGroundwaterLevelPoints = createSelector(
             return [];
         }
         return gwLevels.values.map(data => {
+            const approvalDetails = getDetailsForApprovalCode(data.qualifiers);
             return {
                 value: data.value,
-                dateTime: data.dateTime
+                dateTime: data.dateTime,
+                classes: [GW_LEVEL_CLASS, approvalDetails.class],
+                label: approvalDetails.label,
+                radius: GW_LEVEL_RADIUS
             };
         });
     }
 );
 
 /*
-* When given an approval code, will return the text equivalent
-*  @param {String} approvalCode - Usually a letter such as 'A'
-*   @return {String} - an easy to understand text version of an approval code
-*/
-const approvalCodeText = function(approvalCode) {
-    const approvalText = {
-        P: 'Provisional',
-        A: 'Approved',
-        R: 'Revised',
-        default: `unknown code: ${approvalCode}`
-    };
-
-    return approvalText[approvalCode] || approvalText.default;
-};
+ * Returns a selector function which returns the unique classes/label/radius for the gw level points
+ * @return {Function} which returns an {Array of Object} with the following properties:
+ *      @prop {Array of String} classes
+ *      @prop {String} label
+ *      @prop {Number} radius
+ */
+export const getUniqueGWKinds = createSelector(
+    getGroundwaterLevelPoints,
+    gwPoints => {
+        const allKinds = gwPoints.map(point => {
+            return {
+                classes: point.classes,
+                label: point.label,
+                radius: point.radius
+            };
+        });
+        return uniqWith(allKinds, isEqual);
+    }
+);
 
 /*
  * Selector function which returns a function that returns an array of gw data appropriate
@@ -51,33 +82,25 @@ const approvalCodeText = function(approvalCode) {
  *      @prop {String} parameterName
  *      @prop {String} result
  *      @prop {String} dateTime in site's time zone.
+ *      @prop {String} approvals
  */
 export const getGroundwaterLevelsTableData = createSelector(
-    getGroundwaterLevels,
-    gwLevels => {
-        if (!gwLevels) {
+    getPrimaryParameter,
+    getGroundwaterLevelPoints,
+    (parameter, gwLevels) => {
+        if (!parameter || !gwLevels) {
             return [];
         }
-        return gwLevels.values.map((point) => {
+        return gwLevels.map((point) => {
             return {
-                parameterName: gwLevels.parameter.name,
+                parameterName: parameter.name,
                 result: point.value.toString(),
                 dateTime: DateTime.fromMillis(point.dateTime, {zone: config.locationTimeZone}).toISO({
                     suppressMilliseconds: true,
                     suppressSeconds: true
                 }),
-                approvals: approvalCodeText(point.qualifiers[0])
+                approvals: point.label
             };
         });
     }
-);
-
-/*
- * Returns a selector function that returns true if any ground water
- * levels are visible.
- * @return {Function} which returns {Boolean}
- */
-export const anyVisibleGroundwaterLevels = createSelector(
-    getGroundwaterLevelPoints,
-    (gwLevels) => gwLevels.length !== 0
 );